[Refactor] setting redux-saga
