+++ package.json
@@ -19,6 +19,7 @@
     "react-scripts": "4.0.3",
     "redux": "^4.1.0",
     "redux-persist": "^6.0.0",
+    "redux-saga": "^1.1.3",
     "socket.io-client": "^4.1.2",
     "styled-components": "^5.3.0",
     "styled-reset": "^4.3.4",
@@ -58,6 +59,7 @@
     "@types/styled-components": "^5.1.9",
     "@typescript-eslint/eslint-plugin": "^4.23.0",
     "@typescript-eslint/parser": "^4.23.0",
+    "cra-bundle-analyzer": "^0.1.0",
     "eslint": "^7.26.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-prettier": "^8.3.0",
