+++ src/_reducer/index.ts
@@ -2,10 +2,13 @@ import { configureStore } from '@reduxjs/toolkit';
 import storage from 'redux-persist/lib/storage';
 import { combineReducers } from 'redux';
 import { persistReducer } from 'redux-persist';
+import createSagaMiddleware from 'redux-saga';
+import { all } from 'redux-saga/effects';
 
 import user from './user';
 import map from './map';
-import modal from './modal';
+import modal from './modalSlice';
+import watchOpen from './modalSaga';
 import talent from './talent';
 import chattings from './chattings';
 
@@ -25,9 +28,17 @@ const persistConfig = {
 
 const rootReducer = persistReducer(persistConfig, reducers);
 
+const sagaMiddleware = createSagaMiddleware();
+function* rootSaga() {
+  yield all([watchOpen()]);
+}
+
 const store = configureStore({
   reducer: rootReducer,
+  middleware: [sagaMiddleware],
 });
+
+sagaMiddleware.run(rootSaga);
 
 export type RootState = ReturnType<typeof store.getState>;
 export default store;
