+++ src/components/pages/SigninPage/GoogleSignin.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { GoogleLogin } from 'react-google-login';
-import { openModal } from '../../../_reducer/modal';
+import { openModal } from '../../../_reducer/modalSlice';
 import server from '../../../api/index';
 import Signup from './Signup';
 import { signin, UserState } from '../../../_reducer/user';
