+++ src/components/pages/SigninPage/KakaoSignin.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
-import { openModal } from '../../../_reducer/modal';
+import { openModal } from '../../../_reducer/modalSlice';
 import server from '../../../api';
 import Signup from './Signup';
 import { signin, UserState } from '../../../_reducer/user';
