+++ src/components/pages/TalentRegistrationPage/Sections/LocationSearch.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useCallback, useEffect, useRef } from 'react';
 import { useDispatch } from 'react-redux';
 import styled from 'styled-components';
-import { openModal } from '../../../../_reducer/modal';
+import { openModal } from '../../../../_reducer/modalSlice';
 
 const SEARCH = styled.div`
   //border: 1px solid blue;
