+++ src/components/pages/TalentRegistrationPage/Sections/imageUploader.tsx
@@ -1,7 +1,7 @@
 import React, { useState, memo } from 'react';
 import { useDropzone } from 'react-dropzone';
 import { useDispatch } from 'react-redux';
-import { openModal } from '../../../../_reducer/modal';
+import { openModal } from '../../../../_reducer/modalSlice';
 import server from '../../../../api/index';
 import { CONTAINER, IMAGEBIGDIV, IMAGEDIV, IMAGEP, PLUS, SPAN, IMAGESPAN } from './imageUploaderStyle';
